feat(platform): admit private-http provider endpoints behind host policy
A provider definition may now name an http:// baseUrl when the host is
private/loopback-shaped — the self-hosted-model-server case the schema
comment reserved for a reviewed change, and the seam the e2e/docs mock
gateway needs post-#2876 (the retired secretsEnv provider files were the
only mock wiring, and org_providers reads *.yml only).

The schema stays pure (no env reads): it admits the SHAPE, and
reachability is decided at request boundaries — checkProviderHostPolicy
(cloud-metadata hosts always refused; private hosts refused unless the
operator set TALE_ALLOW_PRIVATE_PROVIDER_HOSTS=1) now also guards the
models-endpoint catalog fetch, which additionally threads the private
host into safeFetch's allowlist so its own private-IP gate agrees with
the policy that just passed. The chat direct-call boundary gains the
same check in the chat-tools commit that follows. isPrivateIp moves to
lib/shared/net/ so the schema and the SSRF layers share one recognizer;
safe_fetch re-exports it for its existing consumers. TTS/embedding
provider calls keep failing closed through safeFetch (audited; no
raw-fetch path sees provider URLs).
